+++ cli/upload.ts
@@ -24,13 +24,14 @@ import * as path from 'path';
 import {print} from '../testing/helpers';
 import {printAndExit as printAndExitImpl} from '../testing/helpers';
 import {readFile} from '../testing/helpers';
+import {resolveApiEndpoint} from './helpers';
 import {tryParseSystemError} from './errors';
 import {v4} from 'uuid';
 import {validateMetadata} from './validate';
 
 interface UploadArgs {
   manifestFile: string;
-  codaApiEndpoint: string;
+  apiEndpoint: string;
   notes?: string;
   intermediateOutputDirectory: string;
   timerStrategy: TimerShimStrategy;
@@ -51,7 +52,7 @@ function cleanup(intermediateOutputDirectory: string, logger: Logger) {
 export async function handleUpload({
   intermediateOutputDirectory,
   manifestFile,
-  codaApiEndpoint,
+  apiEndpoint,
   notes,
   timerStrategy,
   apiToken,
@@ -64,9 +65,10 @@ export async function handleUpload({
   }
 
   const manifestDir = path.dirname(manifestFile);
-  const formattedEndpoint = formatEndpoint(codaApiEndpoint);
-  apiToken = assertApiToken(codaApiEndpoint, apiToken);
-  const packId = assertPackId(manifestDir, codaApiEndpoint);
+  apiEndpoint = resolveApiEndpoint(apiEndpoint, manifestDir);
+  const formattedEndpoint = formatEndpoint(apiEndpoint);
+  apiToken = assertApiToken(apiEndpoint, apiToken);
+  const packId = assertPackId(manifestDir, apiEndpoint);
 
   logger.info('Building Pack bundle...');