+++ cli/whoami.ts
@@ -5,17 +5,19 @@ import {formatEndpoint} from './helpers';
 import {getApiKey} from './config_storage';
 import {isResponseError} from '../helpers/external-api/coda';
 import {printAndExit} from '../testing/helpers';
+import {resolveApiEndpoint} from './helpers';
 import {tryParseSystemError} from './errors';
 
 interface WhoamiArgs {
   apiToken?: string;
-  codaApiEndpoint: string;
+  apiEndpoint: string;
 }
 
-export async function handleWhoami({apiToken, codaApiEndpoint}: ArgumentsCamelCase<WhoamiArgs>) {
-  const formattedEndpoint = formatEndpoint(codaApiEndpoint);
+export async function handleWhoami({apiToken, apiEndpoint}: ArgumentsCamelCase<WhoamiArgs>) {
+  apiEndpoint = resolveApiEndpoint(apiEndpoint);
+  const formattedEndpoint = formatEndpoint(apiEndpoint);
   if (!apiToken) {
-    apiToken = getApiKey(codaApiEndpoint);
+    apiToken = getApiKey(apiEndpoint);
     if (!apiToken) {
       return printAndExit('Missing API token. Please run `coda register` to register one.');
     }